backend/requirements.txt
asgiref==3.9.1
Django==5.2.5
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
